Linked users with products

+++ mystore/src/main/java/com/store/mystore/Product/ProductControllerRestless.java
@@ -5,10 +5,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.criteria.CriteriaBuilder;
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
+import java.io.IOException;
 import java.util.List;
-import java.util.Map;
+
+import com.store.mystore.security.*;
+import org.springframework.web.servlet.view.RedirectView;
 
 @Controller
 @RequestMapping(path="products/") // prefix all urls with `products/`
@@ -20,45 +23,82 @@ public class ProductControllerRestless {
         this.productServices=productServices;
     }
     @RequestMapping("all_products")
-    public String getProducts(Model model) {
+    public Object getProducts(Model model, HttpServletRequest request) throws IOException {
+        if (!Protection.isValidSignin(request.getCookies(), jwt_key.getJwtKey())) {
+            return new RedirectView("/users/login");
+        }
+        String user = Protection.returnUser(request.getCookies(), jwt_key.getJwtKey());
+
         List<Product> products = this.productServices.getProducts();
         System.out.println(products);
 
         model.addAttribute("products", products);
+        model.addAttribute("userRequesting", user);
         return "productListing";
     }
 
     @GetMapping("register_product")
-    public String registerProductGet(Model model) {
+    public Object registerProductGet(Model model, HttpServletRequest request) throws IOException {
+        if (!Protection.isValidSignin(request.getCookies(), jwt_key.getJwtKey())) {
+            return new RedirectView("/users/login");
+        }
         Product product = new Product();
         model.addAttribute("product", product);
         return "productRegister";
     }
 
     @PostMapping("register_product")
-    public String registerProduct(Model model, @ModelAttribute("product") Product product){
+    public Object registerProduct(Model model, @ModelAttribute("product") Product product, HttpServletRequest request) throws IOException {
+        if (!Protection.isValidSignin(request.getCookies(), jwt_key.getJwtKey())) {
+            return new RedirectView("/users/login");
+        }
+        String owner = Protection.returnUser(request.getCookies(), jwt_key.getJwtKey());
         try {
-            this.productServices.addProduct(product);
-            return getProducts(model);
+            this.productServices.addProduct(product, owner);
+            return new RedirectView("/products/all_products");
         } catch (Exception e) {
             model.addAttribute("error", e);
             return "fail";
         }
     }
+
+    @GetMapping("my_products")
+    public Object myProductsGet(Model model, HttpServletRequest request) throws IOException {
+        if (!Protection.isValidSignin(request.getCookies(), jwt_key.getJwtKey())) {
+            return new RedirectView("/users/login");
+        }
+        String user = Protection.returnUser(request.getCookies(), jwt_key.getJwtKey());
+        model.addAttribute("products", this.productServices.fetchUserProducts(user));
+        return "usersProducts";
+    }
+
     @GetMapping("edit_product")
-    public String editProductGet(Model model) {
-        Product tmp = new Product();
-        model.addAttribute("product", tmp);
+    public Object editProductGet(Model model, HttpServletRequest request) throws IOException {
+        if (!Protection.isValidSignin(request.getCookies(), jwt_key.getJwtKey())) {
+            return new RedirectView("/users/login");
+        }
+        String user = Protection.returnUser(request.getCookies(), jwt_key.getJwtKey());
+        List<Product> userProducts = this.productServices.fetchUserProducts(user);
+        model.addAttribute("product", new Product());
+        model.addAttribute("products", userProducts);
         return "editProduct";
     }
 
     @PostMapping("edit_product")
     @Transactional
-    public String editProduct(@ModelAttribute("productPrice") Double newPrice, @ModelAttribute("id") long id, Model model) {
+    public Object editProduct(@ModelAttribute("productPrice") Double newPrice, @ModelAttribute("id") long id, Model model
+                                , HttpServletRequest request) throws IOException {
+        if (!Protection.isValidSignin(request.getCookies(), jwt_key.getJwtKey())) {
+            return new RedirectView("/users/login");
+        }
+        String user = Protection.returnUser(request.getCookies(), jwt_key.getJwtKey());
 
-        String r = this.productServices.updatePrice(id, newPrice);
-
-        return getProducts(model);
+        String r = this.productServices.updatePrice(id, newPrice, user);
+        if (!r.equals("Updated")) {
+            model.addAttribute("error", r);
+            return "fail";
+        }
+        return getProducts(model, request);
     }
 
     @GetMapping("delete_product")
@@ -68,8 +108,11 @@ public class ProductControllerRestless {
     }
 
     @PostMapping("delete_product")
-    public String deleteProduct(@ModelAttribute("id") long id, Model model) {
+    public Object deleteProduct(@ModelAttribute("id") long id, Model model, HttpServletRequest request) throws IOException {
+        if (!Protection.isValidSignin(request.getCookies(), jwt_key.getJwtKey())) {
+            return new RedirectView("/users/login");
+        }
         this.productServices.deleteProduct(id);
-        return getProducts(model);
+        return getProducts(model, request);
     }
 }
